thrust-macros: split lib.rs into context, invariant, and spec modules
Pure refactor: move the existing `context` impl into `context.rs`, and the
`predicate`/`requires`/`ensures`/`_requires_ensures` machinery (plus
`FnItemWithSignature`, `ExpandedTokens`, and their helpers) into `spec.rs`.

The crate root now holds only the proc-macro entry points (thin delegations)
and the two helpers shared across modules (`FnOuterItem`,
`fn_params_with_model_ty`). The shared helpers stay private: a private root
item is visible to every descendant module, so no `pub(crate)` is required.

No behavior change.
